+++ src/GUI/__init__.py
@@ -1,6 +1,6 @@
 from flask import Flask, render_template, request
 
-from src.convertion.convert import convert_to_txt
+from src.conversion.convert import convert_to_txt
 from src.messaging.rabbitmq import publish_message
 
 app = Flask(__name__)
